requirements.txt
customtkinter
cryptography
platformdirs
pillow